fix parsing of an empty style or script tag
This wizardry is inspired by this Discord thread: https://discord.com/channels/1063097320771698699/1126899360995737722

I don't fully understand the internals of tree-sitter, but based on this
thread, I think _before_ this change the parser would only ever consider
the sequence '<', 'style' to be valid for `style_element`.

Now that `element_identifier` explicitly depends on the 'style' token, I
guess the parser "knows" it can recover to the `self_closing_tag` rule.